Move helper functions into correct files
