+++ internal/cmd/db_connect.go
@@ -1,14 +1,23 @@
 package cmd
 
 import (
+	"context"
+	"errors"
 	"fmt"
+	"io"
+	"net/http"
 	"os"
 	"os/exec"
+	"time"
 
+	tea "github.com/charmbracelet/bubbletea"
+	"github.com/jackc/pgx/v5"
+	"github.com/jackc/pgx/v5/pgconn"
 	"github.com/spf13/cobra"
 
 	"github.com/timescale/tiger-cli/internal/api"
 	"github.com/timescale/tiger-cli/internal/common"
+	"github.com/timescale/tiger-cli/internal/util"
 )
 
 func buildDbConnectCmd() *cobra.Command {
@@ -153,6 +162,452 @@ func separateServiceAndPsqlArgs(cmd ArgsLenAtDashProvider, args []string) ([]str
 	}
 
 	return serviceArgs, psqlFlags
+}
+
+// selectConnection returns the connection details for `tiger db connect`. A
+// replica target connects straight through; a primary in an interactive
+// terminal is offered a menu to pick the primary or one of its replicas (nil
+// details means the user cancelled).
+func selectConnection(
+	ctx context.Context,
+	cmd *cobra.Command,
+	client *api.ClientWithResponses,
+	projectID string,
+	target *common.ConnectionTarget,
+	opts common.ConnectionDetailsOptions,
+	noReplicaPrompt bool,
+) (*common.ConnectionDetails, error) {
+	// chosen is what we connect to; the menu below may replace it with a replica.
+	chosen := target
+
+	// Offer the replica menu only for a primary on an interactive terminal.
+	if !target.IsReplica && !noReplicaPrompt && checkStdinIsTTY() {
+		primary := target.ConnectionService
+		replicas, err := fetchReplicaSets(ctx, client, projectID, util.DerefStr(primary.ServiceId))
+		if err != nil {
+			// Don't block the connection if we can't list replicas.
+			fmt.Fprintf(cmd.ErrOrStderr(), "Warning: could not list read replicas: %v\n", err)
+		} else if connectable := connectableReplicas(replicas); len(connectable) > 0 {
+			choice, err := selectConnectTargetOption(cmd.ErrOrStderr(), primary, connectable)
+			if err != nil {
+				return nil, err
+			}
+			switch choice.kind {
+			case targetCancel:
+				return nil, nil
+			case targetReplica:
+				chosen = common.NewReplicaConnectionTarget(primary, *choice.replica)
+			}
+		}
+	}
+
+	details, err := buildConnectionDetailsForTarget(cmd, chosen, opts)
+	if err != nil {
+		return nil, err
+	}
+	if chosen.IsReplica {
+		fmt.Fprintf(cmd.ErrOrStderr(), "Connecting to read replica '%s'...\n", util.DerefStr(chosen.ConnectionService.Name))
+	}
+	return details, nil
+}
+
+// connectableReplicas filters to active read replicas that expose an endpoint.
+func connectableReplicas(replicas []api.ReadReplicaSet) []api.ReadReplicaSet {
+	var out []api.ReadReplicaSet
+	for _, r := range replicas {
+		if r.Status != nil && *r.Status == api.ReadReplicaSetStatusActive && r.Endpoint != nil {
+			out = append(out, r)
+		}
+	}
+	return out
+}
+
+// fetchReplicaSets retrieves the read replica sets for a service.
+func fetchReplicaSets(ctx context.Context, client *api.ClientWithResponses, projectID, serviceID string) ([]api.ReadReplicaSet, error) {
+	ctx, cancel := context.WithTimeout(ctx, 30*time.Second)
+	defer cancel()
+
+	resp, err := client.GetReplicaSetsWithResponse(ctx, projectID, serviceID)
+	if err != nil {
+		return nil, err
+	}
+
+	if resp.StatusCode() != http.StatusOK {
+		return nil, common.ExitWithErrorFromStatusCode(resp.StatusCode(), resp.JSON4XX)
+	}
+
+	if resp.JSON200 == nil {
+		return nil, nil
+	}
+
+	return *resp.JSON200, nil
+}
+
+// connectTargetKind enumerates the choices in the connect target menu.
+type connectTargetKind int
+
+const (
+	targetPrimary connectTargetKind = iota
+	targetReplica
+	targetCancel
+)
+
+// connectTargetChoice is one menu entry: its display label and the action it represents.
+type connectTargetChoice struct {
+	kind    connectTargetKind
+	replica *api.ReadReplicaSet
+	label   string
+}
+
+// connectTargetModel is the Bubble Tea model for selecting a connection target.
+type connectTargetModel struct {
+	choices []connectTargetChoice
+	cursor  int
+	chosen  connectTargetChoice
+}
+
+func newConnectTargetModel(primary api.Service, replicas []api.ReadReplicaSet) connectTargetModel {
+	choices := []connectTargetChoice{{
+		kind:  targetPrimary,
+		label: fmt.Sprintf("Connect to primary service (%s)", util.DerefStr(primary.ServiceId)),
+	}}
+
+	for i := range replicas {
+		choices = append(choices, connectTargetChoice{
+			kind:    targetReplica,
+			replica: &replicas[i],
+			label:   fmt.Sprintf("Connect to read replica '%s'", util.DerefStr(replicas[i].Name)),
+		})
+	}
+
+	choices = append(choices, connectTargetChoice{kind: targetCancel, label: "Cancel"})
+
+	return connectTargetModel{
+		choices: choices,
+		// Default to cancel so quitting (ctrl+c/q) is a no-op connection.
+		chosen: connectTargetChoice{kind: targetCancel},
+	}
+}
+
+func (m connectTargetModel) Init() tea.Cmd {
+	return nil
+}
+
+func (m connectTargetModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
+	switch msg := msg.(type) {
+	case tea.KeyMsg:
+		switch key := msg.String(); key {
+		case "ctrl+c", "q":
+			m.chosen = connectTargetChoice{kind: targetCancel}
+			return m, tea.Quit
+		case "up", "k":
+			if m.cursor > 0 {
+				m.cursor--
+			}
+		case "down", "j":
+			if m.cursor < len(m.choices)-1 {
+				m.cursor++
+			}
+		case "enter", " ":
+			m.chosen = m.choices[m.cursor]
+			return m, tea.Quit
+		default:
+			// Number keys jump straight to that option ('1' -> first, etc.).
+			if len(key) == 1 && key[0] >= '1' && key[0] <= '9' {
+				if idx := int(key[0] - '1'); idx < len(m.choices) {
+					m.cursor = idx
+					m.chosen = m.choices[idx]
+					return m, tea.Quit
+				}
+			}
+		}
+	}
+	return m, nil
+}
+
+func (m connectTargetModel) View() string {
+	s := "How would you like to connect?\n\n"
+
+	for i, choice := range m.choices {
+		cursor := " "
+		if m.cursor == i {
+			cursor = ">"
+		}
+		s += fmt.Sprintf("%s %d. %s\n", cursor, i+1, choice.label)
+	}
+
+	s += "\nUse ↑/↓ arrows or number keys to select, enter to confirm, q to cancel"
+	return s
+}
+
+// selectConnectTargetOption shows the interactive menu for choosing a
+// connection target.
+func selectConnectTargetOption(out io.Writer, primary api.Service, replicas []api.ReadReplicaSet) (connectTargetChoice, error) {
+	model := newConnectTargetModel(primary, replicas)
+
+	program := tea.NewProgram(model, tea.WithOutput(out))
+	finalModel, err := program.Run()
+	if err != nil {
+		return connectTargetChoice{kind: targetCancel}, fmt.Errorf("failed to run connect menu: %w", err)
+	}
+
+	return finalModel.(connectTargetModel).chosen, nil
+}
+
+// connectWithPasswordMenu handles the connection flow if the stored password is invalid
+// Offers an interactive menu to enter the password manually or reset it
+func connectWithPasswordMenu(
+	ctx context.Context,
+	cmd *cobra.Command,
+	client *api.ClientWithResponses,
+	service api.Service,
+	details *common.ConnectionDetails,
+	psqlPath string,
+	psqlFlags []string,
+) error {
+	// Interactive mode: Get stored password (if any)
+	storage := common.GetPasswordStorage()
+	storedPassword, err := storage.Get(service, details.Role)
+	if err != nil {
+		fmt.Fprintf(cmd.ErrOrStderr(), "Warning: could not retrieve stored password: %v\n", err)
+	}
+
+	// Try to connect with stored password first
+	err = testConnectionWithPassword(ctx, details, storedPassword)
+	if err == nil {
+		// Password works, launch psql
+		return launchPsql(details, psqlPath, psqlFlags, service, cmd)
+	}
+
+	// Check if it's an auth error
+	if !isAuthenticationError(err) {
+		// Non-auth error (network, timeout, etc.) - report it directly
+		return err
+	}
+	// Auth failed with stored password, continue to recovery menu
+	fmt.Fprintf(cmd.ErrOrStderr(), "%s\nStored password is likely invalid or expired.\n\n", err.Error())
+
+	// Check if we're in a TTY for interactive menu
+	if !checkStdinIsTTY() {
+		return fmt.Errorf("authentication failed and no TTY available for interactive password entry")
+	}
+
+	// Interactive recovery loop
+	// Only allow password reset for admin role
+	canResetPassword := details.Role == "tsdbadmin"
+	for {
+		option, err := selectPasswordRecoveryOption(cmd.ErrOrStderr(), canResetPassword)
+		if err != nil {
+			return err
+		}
+
+		switch option {
+		case optionEnterPassword:
+			// Prompt for password
+			fmt.Fprint(cmd.ErrOrStderr(), "Enter password: ")
+			password, err := readString(ctx, readPasswordFromTerminal)
+			fmt.Fprintln(cmd.ErrOrStderr()) // newline after password entry
+			if err != nil {
+				if errors.Is(err, context.Canceled) {
+					return nil // user cancelled
+				}
+				fmt.Fprintf(cmd.ErrOrStderr(), "Error reading password: %v\n\n", err)
+				continue
+			}
+
+			// Test, save, and launch
+			details.Password = password
+			if err = testSaveAndLaunchPsqlWithPassword(ctx, cmd, details, psqlPath, psqlFlags, service); err != nil {
+				if isAuthenticationError(err) {
+					fmt.Fprintf(cmd.ErrOrStderr(), "Password incorrect. Please try again.\n\n")
+					continue
+				}
+				return fmt.Errorf("connection failed: %w", err)
+			}
+			return nil
+
+		case optionResetPassword:
+			// Prompt and reset
+			password, err := promptAndResetPassword(ctx, cmd.ErrOrStderr(), client, service, details.Role)
+			if err != nil {
+				if errors.Is(err, context.Canceled) {
+					return nil // user cancelled
+				}
+				fmt.Fprintf(cmd.ErrOrStderr(), "Error resetting password: %v\n\n", err)
+				continue
+			}
+			fmt.Fprintf(cmd.ErrOrStderr(), "✅ Master password for '%s' user updated successfully\n", details.Role)
+			// Launch psql (password is now in storage)
+			details.Password = password
+			return launchPsql(details, psqlPath, psqlFlags, service, cmd)
+
+		case optionExit:
+			return nil
+		}
+	}
+}
+
+// testConnectionWithPassword tests database connectivity with a specific password
+// Returns nil on success, error on failure
+func testConnectionWithPassword(ctx context.Context, details *common.ConnectionDetails, password string) error {
+	// copy details with provided password
+	copyDetails := *details
+	copyDetails.Password = password
+	connStr := copyDetails.String()
+
+	ctx, cancel := context.WithTimeout(ctx, 10*time.Second)
+	defer cancel()
+
+	conn, err := pgx.Connect(ctx, connStr)
+	if err != nil {
+		return err
+	}
+	return conn.Close(ctx)
+}
+
+// isAuthenticationError checks if the error is a PostgreSQL authentication failure
+func isAuthenticationError(err error) bool {
+	if err == nil {
+		return false
+	}
+	// Check for PostgreSQL error code 28P01 (invalid_password) or 28000 (invalid_authorization_specification)
+	var pgErr *pgconn.PgError
+	if errors.As(err, &pgErr) {
+		return pgErr.Code == "28P01" || pgErr.Code == "28000"
+	}
+	return false
+}
+
+// passwordRecoveryOption represents the user's choice in the password recovery menu
+type passwordRecoveryOption int
+
+const (
+	optionEnterPassword passwordRecoveryOption = iota
+	optionResetPassword
+	optionExit
+)
+
+// passwordRecoveryModel is the Bubble Tea model for password recovery selection
+type passwordRecoveryModel struct {
+	options          []string
+	optionMap        []passwordRecoveryOption // maps cursor position to option enum
+	cursor           int
+	selected         passwordRecoveryOption
+	canResetPassword bool
+}
+
+func newPasswordRecoveryModel(canResetPassword bool) passwordRecoveryModel {
+	options := []string{"Enter password manually"}
+	optionMap := []passwordRecoveryOption{optionEnterPassword}
+
+	if canResetPassword {
+		options = append(options, "Update/reset password")
+		optionMap = append(optionMap, optionResetPassword)
+	}
+
+	options = append(options, "Exit")
+	optionMap = append(optionMap, optionExit)
+
+	return passwordRecoveryModel{
+		options:          options,
+		optionMap:        optionMap,
+		cursor:           0,
+		canResetPassword: canResetPassword,
+	}
+}
+
+func (m passwordRecoveryModel) Init() tea.Cmd {
+	return nil
+}
+
+func (m passwordRecoveryModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
+	switch msg := msg.(type) {
+	case tea.KeyMsg:
+		switch msg.String() {
+		case "ctrl+c", "q":
+			m.selected = optionExit
+			return m, tea.Quit
+		case "up", "k":
+			if m.cursor > 0 {
+				m.cursor--
+			}
+		case "down", "j":
+			if m.cursor < len(m.options)-1 {
+				m.cursor++
+			}
+		case "enter", " ":
+			m.selected = m.optionMap[m.cursor]
+			return m, tea.Quit
+		default:
+			// Handle number keys based on available options
+			if len(msg.String()) == 1 && msg.String()[0] >= '1' && msg.String()[0] <= '9' {
+				idx := int(msg.String()[0] - '1') // '1' -> 0, '2' -> 1, etc.
+				if idx >= 0 && idx < len(m.options) {
+					m.cursor = idx
+					m.selected = m.optionMap[idx]
+					return m, tea.Quit
+				}
+			}
+		}
+	}
+	return m, nil
+}
+
+func (m passwordRecoveryModel) View() string {
+	s := "What would you like to do?\n\n"
+
+	for i, option := range m.options {
+		cursor := " "
+		if m.cursor == i {
+			cursor = ">"
+		}
+		s += fmt.Sprintf("%s %d. %s\n", cursor, i+1, option)
+	}
+
+	s += "\nUse ↑/↓ arrows or number keys to select, enter to confirm, q to quit"
+	return s
+}
+
+// selectPasswordRecoveryOption shows the interactive menu for password recovery
+// canResetPassword controls whether the "Update/reset password" option is shown
+func selectPasswordRecoveryOption(out io.Writer, canResetPassword bool) (passwordRecoveryOption, error) {
+	model := newPasswordRecoveryModel(canResetPassword)
+
+	program := tea.NewProgram(model, tea.WithOutput(out))
+	finalModel, err := program.Run()
+	if err != nil {
+		return optionExit, fmt.Errorf("failed to run password recovery menu: %w", err)
+	}
+
+	result := finalModel.(passwordRecoveryModel)
+	return result.selected, nil
+}
+
+// testSaveAndLaunchPsqlWithPassword tests a password, saves it if valid, and launches psql.
+// Returns a retryable error if authentication fails, or a fatal error otherwise.
+func testSaveAndLaunchPsqlWithPassword(
+	ctx context.Context,
+	cmd *cobra.Command,
+	details *common.ConnectionDetails,
+	psqlPath string,
+	psqlFlags []string,
+	service api.Service,
+) error {
+	// Test the password
+	if err := testConnectionWithPassword(ctx, details, details.Password); err != nil {
+		return err
+	}
+
+	// Password works! Save it
+	result, saveErr := common.SavePasswordWithResult(service, details.Password, details.Role)
+	if saveErr != nil {
+		fmt.Fprintf(cmd.ErrOrStderr(), "Warning: could not save password: %v\n", saveErr)
+	} else if result.Success {
+		fmt.Fprintf(cmd.ErrOrStderr(), "%s\n", result.Message)
+	}
+
+	// Launch psql
+	return launchPsql(details, psqlPath, psqlFlags, service, cmd)
 }
 
 // launchPsql launches psql using the connection string and additional flags.
