+++ internal/mcp/utils.go
@@ -47,23 +47,6 @@ func setWithPasswordSchemaProperties(schema *jsonschema.Schema) {
 	schema.Properties["with_password"].Examples = []any{false, true}
 }
 
-// ServiceInfo represents simplified service information for MCP output
-type ServiceInfo struct {
-	ServiceID string        `json:"id" jsonschema:"Service identifier (10-character alphanumeric string)"`
-	Name      string        `json:"name"`
-	Status    string        `json:"status" jsonschema:"Service status (e.g., READY, PAUSED, CONFIGURING, UPGRADING)"`
-	Type      string        `json:"type"`
-	Region    string        `json:"region"`
-	Created   string        `json:"created,omitempty"`
-	Resources *ResourceInfo `json:"resources,omitempty"`
-}
-
-func (ServiceInfo) Schema() *jsonschema.Schema {
-	schema := util.Must(jsonschema.For[ServiceInfo](nil))
-	schema.Properties["type"].Enum = util.AnySlice(validServiceTypes())
-	return schema
-}
-
 // ResourceInfo represents resource allocation information
 type ResourceInfo struct {
 	CPU    string `json:"cpu,omitempty" jsonschema:"CPU allocation (e.g., '0.5 cores', '1 core')"`
@@ -90,51 +73,6 @@ func (ServiceDetail) Schema() *jsonschema.Schema {
 	schema := util.Must(jsonschema.For[ServiceDetail](nil))
 	schema.Properties["type"].Enum = util.AnySlice(validServiceTypes())
 	return schema
-}
-
-// convertToServiceInfo converts an API Service to MCP ServiceInfo
-func (s *Server) convertToServiceInfo(service api.Service) ServiceInfo {
-	info := ServiceInfo{
-		ServiceID: util.Deref(service.ServiceId),
-		Name:      util.Deref(service.Name),
-		Status:    util.DerefStr(service.Status),
-		Type:      util.DerefStr(service.ServiceType),
-		Region:    util.Deref(service.RegionCode),
-	}
-
-	// Add creation time if available
-	if service.Created != nil {
-		info.Created = service.Created.Format("2006-01-02T15:04:05Z")
-	}
-
-	// Add resource information if available
-	if service.Resources != nil && len(*service.Resources) > 0 {
-		resource := (*service.Resources)[0]
-		if resource.Spec != nil {
-			info.Resources = &ResourceInfo{}
-
-			if resource.Spec.CpuMillis != nil {
-				cpuCores := float64(*resource.Spec.CpuMillis) / 1000
-				if cpuCores == float64(int(cpuCores)) {
-					info.Resources.CPU = fmt.Sprintf("%.0f cores", cpuCores)
-				} else {
-					info.Resources.CPU = fmt.Sprintf("%.1f cores", cpuCores)
-				}
-			} else {
-				// CPU is null - this indicates a free tier service
-				info.Resources.CPU = "shared"
-			}
-
-			if resource.Spec.MemoryGbs != nil {
-				info.Resources.Memory = fmt.Sprintf("%d GB", *resource.Spec.MemoryGbs)
-			} else {
-				// Memory is null - this indicates a free tier service
-				info.Resources.Memory = "shared"
-			}
-		}
-	}
-
-	return info
 }
 
 // convertToServiceDetail converts an API Service to MCP ServiceDetail
